requirements.txt
Django
argparse
dj-database-url
dj-static
django-toolbelt
gunicorn
psycopg2
pystache
static
wsgiref
django-static-precompiler
six
django-sampledatahelper
